# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PTSite API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/v1/seasons/{season}/nights/batch": {
            "post": {
                "operationId": "night.storeMany",
                "summary": "Schedule several nights at once, all or none: the dates chosen in the season planner (admins). Each night gets\nthe season's default place. Refuses dates that already have a night in the season",
                "tags": [
                    "Night"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScheduleNightsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Array of `NightResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/NightResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/v1/nights/{night}/reschedule": {
            "post": {
                "operationId": "night.reschedule",
                "summary": "\"Remarcar\": change a scheduled night's date and time. Results keepers and admins",
                "tags": [
                    "Night"
                ],
                "parameters": [
                    {
                        "name": "night",
                        "in": "path",
                        "required": true,
                        "description": "The night ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RescheduleNightRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`NightResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/NightResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "place",
                                                        "results"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "NightResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "season_id": {
                        "type": "integer"
                    },
                    "starts_at": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "scheduled",
                            "open",
                            "finished"
                        ]
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "pot": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "main_event_pot": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "time_chip": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "archived": {
                        "type": "boolean"
                    },
                    "place": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/PlaceResource"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "results": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "position": {
                                    "type": "integer"
                                },
                                "points": {
                                    "type": "string"
                                },
                                "player": {
                                    "$ref": "#/components/schemas/PlayerResource"
                                }
                            },
                            "required": [
                                "position",
                                "points",
                                "player"
                            ]
                        }
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "season_id",
                    "starts_at",
                    "status",
                    "description",
                    "pot",
                    "main_event_pot",
                    "time_chip",
                    "archived",
                    "updated_at"
                ],
                "title": "NightResource"
            },
...
            "PlaceResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "archived": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "address",
                    "archived"
                ],
                "title": "PlaceResource"
            },
...
            "PlayerResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nickname": {
                        "type": "string"
                    },
                    "name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "status": {
                        "type": "string"
                    },
                    "archived": {
                        "type": "boolean"
                    },
                    "thumbnail_version": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Null when the player has no thumbnail. The image is at /players/{id}/thumbnail; add ?v=<this value>\nto the URL, because the answer is cached for a year."
                    },
                    "photo_version": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The same for the larger photo, at /players/{id}/photo."
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Contact details only for admins and the player themself."
                    },
                    "birth_date": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "login": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "description": "Site access, only for admins: null when the player has no login.",
                        "properties": {
                            "username": {
                                "type": "string"
                            },
                            "role": {
                                "type": "string",
                                "enum": [
                                    "player",
                                    "results_keeper",
                                    "admin"
                                ]
                            },
                            "last_login_at": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        },
                        "required": [
                            "username",
                            "role",
                            "last_login_at"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "nickname",
                    "name",
                    "status",
                    "archived",
                    "thumbnail_version",
                    "photo_version",
                    "updated_at"
                ],
                "title": "PlayerResource"
            },
...
            "RescheduleNightRequest": {
                "type": "object",
                "properties": {
                    "starts_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "starts_at"
                ],
                "title": "RescheduleNightRequest"
            },
...
            "ScheduleNightsRequest": {
                "type": "object",
                "properties": {
                    "starts_at": {
                        "type": "array",
                        "description": "Date and time of each night to schedule, at the season's default place.",
                        "items": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "minItems": 1,
                        "maxItems": 100,
                        "uniqueItems": true
                    }
                },
                "required": [
                    "starts_at"
                ],
                "title": "ScheduleNightsRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}